requirements.txt
alabaster==0.7.10
appdirs==1.4.0
asn1crypto==0.22.0
Babel==2.4.0
bleach==2.0.0
braintree==3.37.2
cffi==1.10.0
colorama==0.3.9
cryptography==1.8.1
decorator==4.0.11
Django==1.10.5
django-cors-headers==2.0.2
django-mssql==1.8
django-payments==0.11.0.3
djangorestframework==3.4.6
djangorestframework-expiring-authtoken==0.1.4
docutils==0.13.1
drfdocs==0.0.11
entrypoints==0.2.2
html5lib==0.999999999
idna==2.5
imagesize==0.7.1
ipykernel==4.6.1
ipyparallel==6.0.2
ipython==6.0.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonpickle==0.9.4
jsonschema==2.6.0
jupyter-client==5.0.1
jupyter-core==4.3.0
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.1.1
nbformat==4.3.0
nose==1.3.7
notebook==5.0.0
numpy==1.12.1
packaging==16.8
pandas==0.19.2
pandocfilters==1.4.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
psycopg2==2.7
pycparser==2.17
Pygments==2.2.0
PyJWT==1.5.0
pyparsing==2.1.10
pypiwin32==220
python-dateutil==2.6.0
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.13.0
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.5
stripe==1.55.2
suds-jurko==0.6
testpath==0.3
tornado==4.5.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
wfastcgi==2.2.1
widgetsnbextension==2.0.0
xmltodict==0.11.0
